Update android_lint deps (#23151)
